Fork a first recipe for before/after 9.0
This makes a copy of the recipe into a pre_9.0 folder. The
copy in there uses maxAgp=8.13 and does not use the new boolean
flags introduced in AGP 9.0 (e.g. newDsl, builtInKotlin, etc...)

The old version (in the root recipe folder) is changed to be
minAgp=9.0 but is unchanged otherwise. Moving it to fully 9.0
compatible will be done after all the recipes have been forked.

As part of this change, some minor fixes to the converter was
done to properly handle a forked recipe (since it's the first
time we do this).

Bug: N/A
Test: normal recipe validation
Change-Id: Ib76c1f933f872f09a6c15de3fd0039dc9cbebafd