Fork remaining recipes.
pre 9.0 recipes are in the pre_9.0 folder with a max
agp version of 8.13, and do not use the boolean options
introduced in 9.0

Top level recipes have a min AGP of 9.0.0, and are still
not fully 9.0 compliant, this will come in a later commit.

The testFixture recipe was already 9.0 min and therefore has
not been forked.
selectVariants_pre_8.5 has also not been forked since it
already has a 8.4 max, and was therefore just moved to
the pre_9.0 folder.

Bug: N/A
Test: existing validation.
Change-Id: I9d7f9a982e569b3b1f6e4bef4487b941acf60405