Use macro to find all recipes
Previously we had manually to select every recipe which does
not scale well, especially now that we want to fork all
recipes before/after 9.0

Introduce a macro to find the recipe instead. This also
creates a "recipe_tests" test suite target that runs all the
recipe tests and for each recipe creates a test suite called
"recipe_tests_<name>" that runs all the test for a given
recipe (ie for each version of AGP).

This also fixes a small issue in the GradleRecipeTest when
used with recipes that are not directly inside the recipes
folder. Conversion always uses the last segment of the recipe
location and so additional intermediate folder segments are
skipped. This causes the logic to fail finding converted
recipes.

Bug: N/A
Test: manually (validating that all expected recipes are found)
Change-Id: I0d5863f562af8e01bd304dc43844b2dcc59aecd3