Change how minAgp version is handled in recipe.
Previously the min AGP value accepted any version
scheme, including non-final values (e.g. 9.0.0-alpha03).
This had led to some recipes being stuck on these weird
versions even after the versions went final. This
was cleaned up when the 9.0 recipes were forked from
the pre 9.0 ones.

With 9.0.0 recipes, we also just realized that all
the tests were doing nothing. This is because the
min version of 9.0.0 ends up rejecting versions
like 9.0.0-dev (when running locally), or
9.0.0-alpha12 (when running on CI). While we could
fix it by making all the min versions to be
9.0.0-alpha01, this feels wrong.

This change fixes the issues by changing how we
handle a min version: min version, like max versions
are now using short versions only (x.y).

This is based on the following assumptions:
- we don't change the APIs in patch versions
- difference between alpha versions do not matter
  since we don't have a branch for each alpha
- it's ok for the in development branch to have
  some broken recipe in the <7 days between a ToT
  change and the alpha release that includes it
  (because this is not the default branch on github).

So this commit changes how the version checks are
done for recipes, and makes the recipe metadata only
accept short versions for minAgp.

This also fix the kotlin repo from the test.

Finally, this makes some clean up in the converter
and test harness for the recipes. This attemps to
introduce a way to detect skipped test (due to
AGP version mismatch) to mark them as ignored.

This also simplifies how the tests are run. They
used to create a working copy of the recipe, and
then validate this which would converted it back
to a source copy and then make a release copy from
this. This does not make any sense. We're now
doing a source level validation which just makes
a single release copy (ToT used to _also_ do this).

Bug: N/A
Test: existing recipe tests
Change-Id: Ib040631f894466f49380059d8216a499bbfa1c40